나머지가 1이 되는 수 찾기 문제 풀이

+++ src/test/java/com/ki/practice/p12954/Tests.java
@@ -14,19 +14,19 @@ public class Tests {
     @Test
     @DisplayName("2, 5 => [2,4,6,8,10]")
     void t1() {
-        assertThat(new Solution().solution(2, 5)).isEqualTo(new int[]{2, 4, 6, 8, 10});
+        assertThat(new Solution().solution(2, 5)).isEqualTo(new long[]{2L, 4L, 6L, 8L, 10L});
     }
 
     @Test
     @DisplayName("4, 3 => [4,8,12]")
     void t2() {
-        assertThat(new Solution().solution(4, 3)).isEqualTo(new int[]{4, 8, 12});
+        assertThat(new Solution().solution(4, 3)).isEqualTo(new long[]{4L, 8L, 12L});
     }
 
     @Test
     @DisplayName("-4, 2 => [-4, -8]")
     void t3() {
-        assertThat(new Solution().solution(-4, 2)).isEqualTo(new int[]{-4, -8});
+        assertThat(new Solution().solution(-4, 2)).isEqualTo(new long[]{-4L, -8L});
     }
 
 }
